// src/types/PersonDetailsType.ts
export type PersonDetailsType = {
  currentOrganization?: string;
  summary?: string;
  skills?: string;
  currentEmploymentStatus?: string;
  availableFrom?: string;
  dateOfBirth?: string;
  currentSalary?: string;
  relevantExperience?: string;
  noticePeriod?: string;
  salaryExpectation?: string;
  fullAddress?: string;
  status?: string;
  salaryType?: string;
  resume?: string;
  totalExperience?: string;
  languageSkills?: string;
};